requirements.txt
yfinance>=0.2.30
fredapi>=0.5.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
streamlit>=1.30.0
pyyaml>=6.0
python-dotenv>=1.0.0
scipy>=1.11.0
pyarrow>=14.0.0
pytest>=7.4.0
pytest-watch>=4.2.0
